+++ dev_test/lib/src/package/universal/package_universal_stub.dart
@@ -1,0 +1,4 @@
+Future<Map<String, dynamic>> pathGetPubspecYamlMap(String packageDir) =>
+    throw UnsupportedError('pathGetPubspecYamlMap io only');
+Future<Map<String, dynamic>> pathGetAnalysisOptionsYamlMap(String packageDir) =>
+    throw UnsupportedError('pathGetAnalysisOptionsYamlMap io only');