+++ dev_test/CHANGELOG.md
@@ -1,6 +1,7 @@
-## 0.13.2-dev.3
+## 0.13.3+7
 
-* Add `ioPackageRunCi` to run common dart flutter package test
+* Add `ioPackageRunCi` to run common dart/flutter package test, node test and nnbd tests.
+* Skipped `pub run build_runner test -- -p chrome` on travis for ioPackageRunCi
 
 ## 0.13.1+1